+++ src/components/Menus/Sidebar/Sidebar.jsx
@@ -12,8 +12,9 @@ import HomeIcon from "../../../assets/icons/home.svg";
 import ImageIcon from "../../../assets/icons/image.svg";
 import InfoIcon from "../../../assets/icons/info.svg";
 import SettingsIcon from "../../../assets/icons/settings.svg";
-import DownloadIcon from "../../../assets/icons/download.svg";
 import StepsIcon from "../../../assets/icons/steps.svg";
+import SaveIcon from "../../../assets/icons/save.svg";
+
 import ProjectsPanel from "./ProjectsPanel/ProjectsPanel";
 
 import "../../../assets/stylesheets/Sidebar.scss";
@@ -57,7 +58,7 @@ const Sidebar = ({ options = [] }) => {
     },
     {
       name: "download",
-      icon: DownloadIcon,
+      icon: SaveIcon,
       title: t("sidebar.download"),
       position: "top",
       panel: DownloadPanel,